requirements.txt
Django==2.2.16
djangorestframework==3.12.4
djangorestframework_simplejwt==4.8.0
PyJWT==2.1.0